프로필 페이지, my_page 파일로 묶기 완료.
