+++ lib/ui/components/custom_app_bar.dart
@@ -1,5 +1,5 @@
 import 'package:final_project_team02/ui/holder/cart/cart_page.dart';
-import 'package:final_project_team02/ui/holder/profile/settings_page.dart';
+import 'package:final_project_team02/ui/holder/my_page/pages/settings/settings_page.dart';
 import 'package:flutter/material.dart';
 
 import '../holder/main_holder.dart';
