--- lib/ui/holder/my_page/creator/creator_my_page.dart
@@ -1,30 +1,0 @@
-import 'package:final_project_team02/ui/holder/my_page/creator/components/creator_my_page_header.dart';
-import 'package:flutter/material.dart';
-
-import '../../../components/custom_app_bar.dart';
-import 'components/creator_my_page_bottom.dart';
-
-class CreatorMyPage extends StatelessWidget {
-  @override
-  Widget build(BuildContext context) {
-    return Scaffold(
-      appBar: CustomAppBar(
-        home: Icons.home,
-        setting: Icons.settings,
-        cart: null,
-      ),
-      backgroundColor: Colors.white,
-      body: NestedScrollView(
-        headerSliverBuilder: (BuildContext context, bool innerBoxIsScrolled) {
-          return [
-            SliverToBoxAdapter(
-              //프로필 설정 버튼까지
-              child: CreatorMyPageheader(),
-            ),
-          ];
-        },
-        body: CreatorMyPageBottom(),
-      ),
-    );
-  }
-}
