+++ lib/ui/holder/my_page/pages/user/components/apply/user_my_page_apply_modal_body_form_apply_button.dart
@@ -1,6 +1,5 @@
+import 'package:final_project_team02/ui/holder/home/home_page.dart';
 import 'package:flutter/material.dart';
-
-import '../../../../home/home_page.dart';
 
 class UserMyPageApplyModalBodyApplyButton extends StatelessWidget {
   @override
