+++ lib/ui/holder/my_page/pages/user/components/user_my_page_body_bottom.dart
@@ -1,6 +1,6 @@
-import 'package:final_project_team02/ui/holder/my_page/user/components/user_my_page_body_bottom_tab_bar_closet.dart';
-import 'package:final_project_team02/ui/holder/my_page/user/components/user_my_page_body_bottom_tab_bar_grid.dart';
-import 'package:final_project_team02/ui/holder/my_page/user/components/user_my_page_body_bottom_tab_var.dart';
+import 'package:final_project_team02/ui/holder/my_page/pages/user/components/user_my_page_body_bottom_tab_bar_closet.dart';
+import 'package:final_project_team02/ui/holder/my_page/pages/user/components/user_my_page_body_bottom_tab_bar_grid.dart';
+import 'package:final_project_team02/ui/holder/my_page/pages/user/components/user_my_page_body_bottom_tab_var.dart';
 import 'package:flutter/material.dart';
 import 'package:intl/intl.dart';
